+++ src/Core/class-better-install.php
@@ -20,7 +20,7 @@ class Better_Install {
 	 *
 	 * @since 1.0.0
 	 */
-	const DB_VERSION = '1.1.0';
+	const DB_VERSION = '1.2.0';
 
 	/**
 	 * Create or upgrade import database tables.
@@ -105,6 +105,7 @@ class Better_Install {
 			KEY job_entity (job_id, entity_index),
 			KEY job_status (job_id, status),
 			KEY job_status_step (job_id, status, step),
+			KEY job_status_entity (job_id, status, entity_index),
 			KEY job_type_old (job_id, entity_type, old_entity_id)
 		) {$charset_collate};";